refactor: changed logic useTooltip renderContainer
