+++ ui-parts/tooltip/src/container/container.interfaces.ts
@@ -1,3 +1,13 @@
+import { LayerSide, UseLayerArrowProps } from "react-laag"
+
+export interface ArrowOptions {
+  angle?: number | undefined
+  size?: number | undefined
+  roundness?: number | undefined
+  borderWidth?: number | undefined
+  borderColor?: string | undefined
+  backgroundColor?: string | undefined
+}
 export interface ContainerElementProps {
   animate?: boolean
   ref?: any
@@ -6,5 +16,8 @@ export interface ContainerElementProps {
 
 export interface DefaultContainerProps extends ContainerElementProps {
   text?: string | number
-  arrow?: JSX.Element
+  showArrow?: boolean
+  layerSide?: LayerSide
+  arrowProps?: UseLayerArrowProps
+  arrowOptions?: ArrowOptions
 }
