+++ ui-parts/tooltip/src/index.ts
@@ -1,3 +1,4 @@
 export * from './tooltip.component'
 export * from './tooltip.interfaces'
+export * from './use-tooltip.hook'
 export { Arrow } from 'react-laag'
